+++ src/main/java/tools/devnull/jenkins/plugins/buildnotifications/PushoverNotifier.java
@@ -33,50 +33,36 @@ import hudson.model.AbstractProject;
 import hudson.model.BuildListener;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.BuildStepMonitor;
-import hudson.tasks.Notifier;
 import hudson.tasks.Publisher;
-import jenkins.model.Jenkins;
 import net.sf.json.JSONObject;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.StaplerRequest;
-
-import java.io.IOException;
 
 /**
  * A notifier that uses Pushover to delivery messages
  *
  * @author Ataxexe
  */
-public class PushoverNotifier extends Notifier {
-
-  private final String userToken;
-  private final String device;
-  private final boolean sendIfSuccess;
+public class PushoverNotifier extends BaseNotifier {
 
   /**
-   * Creates a new notifier based on the given parameters
+   * * Creates a new notifier based on the given parameters
    *
-   * @param userToken     the pushover target token
-   * @param device        the optional device to send the message
-   * @param sendIfSuccess if the notification should be sent if the build succeed
+   * @param globalTarget      the target for all notifications
+   * @param successfulTarget  the target for build success notifications
+   * @param brokenTarget      the target for broken build notifications
+   * @param stillBrokenTarget the target for still broken build notifications
+   * @param fixedTarget       the target for fixed build notifications
+   * @param sendIfSuccess     if the notification should be sent if the build succeed
    */
   @DataBoundConstructor
-  public PushoverNotifier(String userToken, String device, boolean sendIfSuccess) {
-    this.userToken = userToken;
-    this.device = device;
-    this.sendIfSuccess = sendIfSuccess;
-  }
-
-  public String getUserToken() {
-    return userToken;
-  }
-
-  public String getDevice() {
-    return device;
-  }
-
-  public boolean isSendIfSuccess() {
-    return sendIfSuccess;
+  public PushoverNotifier(String globalTarget,
+                          String successfulTarget,
+                          String brokenTarget,
+                          String stillBrokenTarget,
+                          String fixedTarget,
+                          boolean sendIfSuccess) {
+    super(globalTarget, successfulTarget, brokenTarget, stillBrokenTarget, fixedTarget, sendIfSuccess);
   }
 
   @Override
@@ -85,14 +71,9 @@ public class PushoverNotifier extends Notifier {
   }
 
   @Override
-  public boolean perform(AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener)
-      throws InterruptedException, IOException {
-    PushoverDescriptor descriptor = (PushoverDescriptor) Jenkins.getInstance()
-        .getDescriptor(PushoverNotifier.class);
-    Message message = new PushoverMessage(userToken, descriptor.appToken, device);
-    BuildNotifier notifier = new BuildNotifier(message, build, sendIfSuccess);
-    notifier.sendNotification();
-    return true;
+  protected Message createMessage(String target, AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener) {
+    PushoverDescriptor descriptor = (PushoverDescriptor) getDescriptor();
+    return new PushoverMessage(target, descriptor.appToken);
   }
 
   /**